Refactor monster book data
